Add README.md
Added a `README.md` file. Includes details about the project, why the project was created, stubs for how to run the project yourself, a new screenshot, and a simple list of features.

Note - there are a few `TODO` statements which should eventually be filled out by someone more familiar with the project structure.